requirements.txt
opencv-python
pillow
torch
transformers
deep_translator